refactor(api): implement three-layer API service architecture
Layer 1 (API Service): ZhjwApiService, WfwApiService, PayAppApiService, CcylApiService
Layer 2 (Subsystem Auth): ZhjwAuth, WfwAuth, PayAppAuth, FitnessAuth, CcylAuth
Layer 3 (Unified Auth): ScuAuth (merged ScuAuthService + ScuAuthSession)

- Add unified exception hierarchy: UnauthenticatedException, ServiceException, ScuLoginException
- Remove AuthManager; layers depend directly on each other
- Move HTTP + HTML parsing logic from TrainProgramProvider/PlanCompletionProvider into ZhjwApiService
- Slim down providers to pure UI state management
- Each layer provides auto-retry via _request() with UnauthenticatedException bubbling upward
- ScuAuth has built-in _synchronizedRefresh mutex for concurrent request safety
