+++ lib/pages/campus/classroom/classroom_page.dart
@@ -4,7 +4,8 @@ import 'package:bugaoshan/l10n/app_localizations.dart';
 import 'package:bugaoshan/pages/campus/classroom/classroom_detail_page.dart';
 import 'package:bugaoshan/pages/campus/models/classroom_model.dart';
 import 'package:bugaoshan/providers/scu_auth_provider.dart';
-import 'package:bugaoshan/services/scu_api_service.dart';
+import 'package:bugaoshan/services/api/zhjw_api_service.dart';
+import 'package:bugaoshan/services/exceptions/scu_exceptions.dart';
 import 'package:bugaoshan/widgets/common/loading_widgets.dart';
 import 'package:bugaoshan/widgets/common/login_required_widget.dart';
 import 'package:bugaoshan/widgets/common/error_widgets.dart';
@@ -19,7 +20,7 @@ class ClassroomPage extends StatefulWidget {
 }
 
 class _ClassroomPageState extends State<ClassroomPage> {
-  late final ScuApiService _authService;
+  late final ZhjwApiService _zhjwApi;
 
   List<ClassroomCampus> _campuses = [];
   List<ClassroomBuilding> _allBuildings = [];
@@ -36,7 +37,7 @@ class _ClassroomPageState extends State<ClassroomPage> {
   @override
   void initState() {
     super.initState();
-    _authService = getIt<ScuAuthProvider>().service;
+    _zhjwApi = getIt<ZhjwApiService>();
     getIt<ScuAuthProvider>().addListener(_onAuthChanged);
     _loadIndex();
   }
@@ -74,7 +75,7 @@ class _ClassroomPageState extends State<ClassroomPage> {
       _error = null;
     });
     try {
-      final result = await _authService.fetchClassroomIndex();
+      final result = await _zhjwApi.fetchClassroomIndex();
       if (!mounted) return;
       setState(() {
         _campuses = result.campuses;
@@ -82,10 +83,10 @@ class _ClassroomPageState extends State<ClassroomPage> {
         _isLoading = false;
         _isInitialLoad = false;
       });
-    } on ScuLoginException catch (e) {
+    } on UnauthenticatedException catch (_) {
       if (!mounted) return;
       setState(() {
-        _error = e.sessionExpired ? 'sessionExpired' : 'loadFailed';
+        _error = 'sessionExpired';
         _isLoading = false;
         _isInitialLoad = false;
       });
@@ -120,7 +121,7 @@ class _ClassroomPageState extends State<ClassroomPage> {
     try {
       final dateStr =
           '${_selectedDate.year}-${_selectedDate.month.toString().padLeft(2, '0')}-${_selectedDate.day.toString().padLeft(2, '0')}';
-      _queryResult = await _authService.fetchClassroomAvailability(
+      _queryResult = await _zhjwApi.fetchClassroomAvailability(
         campusNumber: building.campusNumber,
         buildingNumber: building.teachingBuildingNumber,
         searchDate: dateStr,
@@ -129,10 +130,10 @@ class _ClassroomPageState extends State<ClassroomPage> {
       setState(() {
         _isLoading = false;
       });
-    } on ScuLoginException catch (e) {
+    } on UnauthenticatedException catch (_) {
       if (!mounted) return;
       setState(() {
-        _error = e.sessionExpired ? 'sessionExpired' : 'loadFailed';
+        _error = 'sessionExpired';
         _isLoading = false;
       });
     } catch (e) {
