+++ lib/providers/grades_provider.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:bugaoshan/models/scheme_score.dart';
-import 'package:bugaoshan/providers/scu_auth_provider.dart';
-import 'package:bugaoshan/services/scu_api_service.dart';
+import 'package:bugaoshan/services/api/zhjw_api_service.dart';
+import 'package:bugaoshan/services/exceptions/scu_exceptions.dart';
 
 const _keySchemeScores = 'grades_scheme_scores';
 const _keyPassingScores = 'grades_passing_scores';
@@ -12,9 +12,9 @@ enum GradesLoadState { idle, loading, loaded, error }
 
 class GradesProvider extends ChangeNotifier {
   final SharedPreferences _prefs;
-  final ScuAuthProvider _authProvider;
+  final ZhjwApiService _zhjwApi;
 
-  GradesProvider(this._prefs, this._authProvider) {
+  GradesProvider(this._prefs, this._zhjwApi) {
     final cachedScheme = _prefs.getString(_keySchemeScores);
     if (cachedScheme != null) {
       try {
@@ -50,11 +50,11 @@ class GradesProvider extends ChangeNotifier {
     _schemeError = null;
     notifyListeners();
     try {
-      final data = await _authProvider.service.fetchSchemeScores();
+      final data = await _zhjwApi.fetchSchemeScores();
       _schemeScores = SchemeScoreSummary.fromJson(data);
       _schemeState = GradesLoadState.loaded;
       await _prefs.setString(_keySchemeScores, jsonEncode(data));
-    } on ScuLoginException {
+    } on UnauthenticatedException {
       if (_schemeScores != null) {
         _schemeState = GradesLoadState.loaded;
         _schemeError = 'sessionExpired';
@@ -94,11 +94,11 @@ class GradesProvider extends ChangeNotifier {
     _passingError = null;
     notifyListeners();
     try {
-      final data = await _authProvider.service.fetchPassingScores();
+      final data = await _zhjwApi.fetchPassingScores();
       _passingScores = PassingScoreResult.fromJson(data);
       _passingState = GradesLoadState.loaded;
       await _prefs.setString(_keyPassingScores, jsonEncode(data));
-    } on ScuLoginException {
+    } on UnauthenticatedException {
       if (_passingScores != null) {
         _passingState = GradesLoadState.loaded;
         _passingError = 'sessionExpired';
