+++ lib/services/api/payapp_api_service.dart
@@ -1,0 +1,73 @@
+import 'package:bugaoshan/services/auth/payapp_auth.dart';
+import 'package:bugaoshan/services/balance_query_service.dart';
+import 'package:bugaoshan/services/exceptions/scu_exceptions.dart';
+import 'package:bugaoshan/services/scu_api/cookie_client.dart';
+
+/// 缴费平台 API Service（第1层）
+///
+/// payapp.scu.edu.cn 的业务 API：电费、空调余额查询。
+class PayAppApiService {
+  final PayAppAuth _auth;
+  final BalanceQueryService _service = BalanceQueryService();
+  PayAppApiService(this._auth);
+
+  Future<T> _request<T>(Future<T> Function(CookieClient client) fn) async {
+    try {
+      final client = await _auth.getClient();
+      return await fn(client);
+    } on UnauthenticatedException {
+      final client = await _auth.getClient();
+      return await fn(client);
+    }
+  }
+
+  /// 获取校区列表
+  Future<List<CampusItem>> getCampus() {
+    return _request((client) => _service.getCampus(client));
+  }
+
+  /// 获取建筑列表
+  Future<List<BuildingItem>> getArchitecture(String schoolCode) {
+    return _request((client) => _service.getArchitecture(client, schoolCode));
+  }
+
+  /// 获取单元列表
+  Future<List<UnitItem>> getUnit(String schoolCode, String regCode) {
+    return _request((client) => _service.getUnit(client, schoolCode, regCode));
+  }
+
+  /// 验证/绑定房间
+  Future<bool> verificationRoom({
+    required String cusNo,
+    required int type,
+    required String cusName,
+    required String schoolCode,
+    required String regCode,
+    required String unitCode,
+    required String roomNo,
+  }) {
+    return _request(
+      (client) => _service.verificationRoom(
+        client,
+        cusNo,
+        type,
+        cusName,
+        schoolCode,
+        regCode,
+        unitCode,
+        roomNo,
+      ),
+    );
+  }
+
+  /// 查询房间信息（电费/空调余额）
+  Future<RoomInfo> queryRoomInfo({
+    required String cusNo,
+    required int type,
+    required String cusName,
+  }) {
+    return _request(
+      (client) => _service.queryRoomInfo(client, cusNo, type, cusName),
+    );
+  }
+}
