+++ lib/services/ccyl_oauth_service.dart
@@ -1,16 +1,16 @@
 import 'package:bugaoshan/injection/injector.dart';
-import 'package:bugaoshan/services/auth/auth_manager.dart';
+import 'package:bugaoshan/services/auth/scu_auth.dart';
 import 'package:flutter/widgets.dart';
 import 'package:bugaoshan/utils/constants.dart';
 
 class CcylOAuthService {
   static const _idBase = 'https://id.scu.edu.cn';
   Future<String?> getOAuthCode() async {
-    final authManager = getIt<AuthManager>();
-    final accessToken = authManager.scu.accessToken;
+    final scuAuth = getIt<ScuAuth>();
+    final accessToken = scuAuth.accessToken;
     if (accessToken == null) return null;
 
-    final client = await authManager.scu.getClient();
+    final client = await scuAuth.getClient();
 
     final spLoggedUrl = Uri.parse(
       '$_idBase/api/bff/v1.2/commons/sp_logged'
