--- lib/services/scu_api/scu_api_grades.dart
@@ -1,98 +1,0 @@
-part of 'scu_api_service.dart';
-
-/// 成绩查询相关的教务系统 API。
-extension ScuApiGrades on ScuApiService {
-  /// 获取及格成绩
-  Future<Map<String, dynamic>> fetchPassingScores() async {
-    return _authManager.scu.request((client) async {
-      final indexResp = await client.get(
-        Uri.parse(
-          '$kZhjwBase/student/integratedQuery/scoreQuery/allPassingScores/index',
-        ),
-        headers: {
-          'Accept':
-              'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-          'Referer': '$kZhjwBase/',
-          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
-        },
-      );
-      final indexBody = indexResp.body;
-      _checkSessionExpiry(indexBody, indexResp.statusCode);
-      final urlMatch = RegExp(
-        r'var\s+url\s*=\s*"(/student/integratedQuery/scoreQuery/[^/]+/allPassingScores/callback)"',
-      ).firstMatch(indexBody);
-      if (urlMatch == null) {
-        if (indexBody.contains('login') || indexBody.contains('Login')) {
-          throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-        }
-        throw ScuLoginException('无法从页面提取 allPassingScores callback URL');
-      }
-      final callbackPath = urlMatch.group(1)!;
-
-      final callbackResp = await client.get(
-        Uri.parse('$kZhjwBase$callbackPath'),
-        headers: {
-          'Accept': 'application/json, text/plain, */*',
-          'Referer':
-              '$kZhjwBase/student/integratedQuery/scoreQuery/allPassingScores/index',
-          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
-        },
-      );
-      final body = callbackResp.body.trim();
-      _checkSessionExpiry(body, callbackResp.statusCode);
-      return parseJson(
-        body,
-        'allPassingScores/callback',
-        (msg) => ScuLoginException(msg),
-      );
-    });
-  }
-
-  /// 获取方案成绩
-  Future<Map<String, dynamic>> fetchSchemeScores() async {
-    return _authManager.scu.request((client) async {
-      final indexResp = await client.get(
-        Uri.parse(
-          '$kZhjwBase/student/integratedQuery/scoreQuery/schemeScores/index',
-        ),
-        headers: {
-          'Accept':
-              'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-          'Referer': '$kZhjwBase/',
-          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
-        },
-      );
-
-      final indexBody = indexResp.body;
-      _checkSessionExpiry(indexBody, indexResp.statusCode);
-
-      final urlMatch = RegExp(
-        r'var\s+url\s*=\s*"(/student/integratedQuery/scoreQuery/[^/]+/schemeScores/callback)"',
-      ).firstMatch(indexBody);
-      if (urlMatch == null) {
-        if (indexBody.contains('login') || indexBody.contains('Login')) {
-          throw ScuLoginException('登录已过期，请重新登录', sessionExpired: true);
-        }
-        throw ScuLoginException('无法从页面提取 schemeScores callback URL');
-      }
-      final callbackPath = urlMatch.group(1)!;
-      final callbackResp = await client.get(
-        Uri.parse('$kZhjwBase$callbackPath'),
-        headers: {
-          'Accept': 'application/json, text/plain, */*',
-          'Referer':
-              '$kZhjwBase/student/integratedQuery/scoreQuery/schemeScores/index',
-          'User-Agent': ScuAuthService.requestHeaders['User-Agent']!,
-        },
-      );
-
-      final body = callbackResp.body.trim();
-      _checkSessionExpiry(body, callbackResp.statusCode);
-      return parseJson(
-        body,
-        'schemeScores/callback',
-        (msg) => ScuLoginException(msg),
-      );
-    });
-  }
-}
